Move QrInvite out to a submodle
